# packages/shared/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Norbert's Spark API",
    "description": "API for Norbert's Spark AI CRM application",
    "version": "1.0.0",
    "contact": {
      "name": "Norbert's Spark API",
      "email": "api@example.com"
    },
    "license": {
      "name": "AGPLv3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server (HTTP) - API v1"
    },
    {
      "url": "https://127.0.0.1:3001/api/v1",
      "description": "Development server (HTTPS) - API v1"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server (HTTP) - Root"
    },
    {
      "url": "https://127.0.0.1:3001",
      "description": "Development server (HTTPS) - Root"
    }
  ],
...
  "paths": {
...
    "/ai/chats/{userId}": {
      "get": {
        "summary": "Get chat IDs for user",
        "description": "Retrieves a list of chat IDs associated with the specified user",
        "operationId": "getAIChatsByUserId",
        "tags": ["ai", "chat"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The user's unique identifier whose chat history is being retrieved",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "01890c3a-6f2b-7c1a-b9e1-9b5a0d5f6e3a"
          }
        ],
        "responses": {
          "200": {
            "description": "AI response found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIUserIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidUserId": {
                    "summary": "Invalid userId parameter",
                    "value": {
                      "success": false,
                      "error": "Invalid userId parameter"
                    }
                  },
                  "invalidFormat": {
                    "summary": "Invalid userId format",
                    "value": {
                      "success": false,
                      "error": "Invalid userId format"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required role (user, moderator or admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Access denied. You can only access your own chat history or must have admin/moderator role"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
    "/ai/extract-data/presigned-urls": {
      "post": {
        "summary": "Generate presigned URLs for direct R2 upload",
        "description": "Accepts file metadata (filename and mimetype) and returns presigned URLs for direct upload to Cloudflare R2 storage. The client can then use these URLs to upload files directly without going through the backend.",
        "operationId": "generatePresignedUrls",
        "tags": ["ai", "data"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "description": "Array of file metadata objects",
                    "items": {
                      "type": "object",
                      "properties": {
                        "filename": {
                          "type": "string",
                          "description": "Name of the file",
                          "example": "document.pdf"
                        },
                        "mimetype": {
                          "type": "string",
                          "description": "MIME type of the file",
                          "example": "application/pdf"
                        }
                      },
                      "required": ["filename", "mimetype"]
                    }
                  }
                },
                "required": ["files"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URLs generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "uploadUrls": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "filename": {
                                "type": "string",
                                "description": "Original filename",
                                "example": "document.pdf"
                              },
                              "uploadUrl": {
                                "type": "string",
                                "description": "Presigned URL for direct upload to R2",
                                "example": "https://r2.example.com/bucket/data-extraction/uuid/document.pdf?signature=..."
                              },
                              "fileKey": {
                                "type": "string",
                                "description": "Storage key for the file in R2",
                                "example": "data-extraction/01HXYZ123/document.pdf"
                              }
                            },
                            "required": ["filename", "uploadUrl", "fileKey"]
                          }
                        }
                      },
                      "required": ["uploadUrls"]
                    },
                    "message": {
                      "type": "string",
                      "example": "Presigned URLs generated successfully"
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity - Invalid file type or missing file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ai/extract-data/{fileId}": {
      "get": {
        "summary": "Fetch data extracted from PDFs",
        "description": "Retrieves the data from the PDF file previously uploaded to bucket",
        "operationId": "extractData",
        "tags": ["ai", "chat"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "description": "The unique identifier for the file in the bucket",
            "schema": { "type": "string", "format": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Streaming text response with extracted data",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Streamed text response containing the extracted data from the PDF"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required role (user, moderator or admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Access denied. You can only access your own extracted data or must have admin/moderator role"
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Extracted data not found for the specified file"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AIUserIdResponse": {
        "type": "object",
        "description": "Structured response containing array of chat IDs for a user",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the request was successful",
            "example": true
          },
          "data": {
            "type": "array",
            "description": "Array of UUID strings for previous AI chats",
            "items": {
              "type": "string",
              "format": "uuid",
              "example": "01890c3a-6f2b-7c1a-b9e1-9b5a0d5f6e3a"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "Standard error response object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid request data"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated requests"
      },
      "oauthSyncSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OAuth-Sync-Secret",
        "description": "Shared secret for OAuth sync endpoint authentication. Must match OAUTH_SYNC_SECRET environment variable."
      }
    }
  }
}